# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Incidents",
        "description": "API complète pour la gestion des incidents municipaux",
        "contact": {
            "name": "Support API",
            "email": "support@incidents.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/municipalities/{municipality}": {
            "get": {
                "tags": [
                    "Municipalities"
                ],
                "summary": "Voir une municipalité spécifique",
                "description": "Afficher une municipalité spécifique",
                "operationId": "e0a4cb7edd125343c36aef7cc22b5102",
                "parameters": [
                    {
                        "name": "municipality",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la municipalité"
                    },
                    "404": {
                        "description": "Municipalité non trouvée"
                    }
                }
            }
        },
        "/admin_manager/my-municipality": {
            "get": {
                "tags": [
                    "Municipalities"
                ],
                "summary": "Récupérer ma municipalité",
                "description": "Récupérer ma municipalité (admin municipal ou agent)",
                "operationId": "9f7f27ba3ce506ecce6055a62f0ac91e",
                "responses": {
                    "200": {
                        "description": "Ma municipalité"
                    },
                    "404": {
                        "description": "Municipalité non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/municipalities": {
            "post": {
                "tags": [
                    "Municipalities"
                ],
                "summary": "Créer une municipalité",
                "description": "Créer une municipalité (super admin)",
                "operationId": "05e1e6daf33aef852c390cf8b90c86b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "governorate",
                                    "address",
                                    "city"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Municipalité de Monastir"
                                    },
                                    "governorate": {
                                        "type": "string",
                                        "example": "Monastir"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "522123456",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rue principale",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "monastir@example.com",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Monastir",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Municipalité créée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/municipalities/{municipality}": {
            "put": {
                "tags": [
                    "Municipalities"
                ],
                "summary": "Mettre à jour une municipalité",
                "description": "Mettre à jour une municipalité (super admin)",
                "operationId": "bf078f0bdf9d612b0054f41ffe2357c7",
                "parameters": [
                    {
                        "name": "municipality",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Municipalité mise à jour"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Municipalities"
                ],
                "summary": "Supprimer une municipalité",
                "description": "Supprimer une municipalité (super admin)",
                "operationId": "c8fc6a601ec761a5b3764be01103445e",
                "parameters": [
                    {
                        "name": "municipality",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Municipalité supprimée"
                    },
                    "404": {
                        "description": "Municipalité non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin_manager/municipalities/my-municipality": {
            "put": {
                "tags": [
                    "Municipalities"
                ],
                "summary": "Mettre à jour ma municipalité (Admin Manager)",
                "description": "Mettre à jour ma municipalité (admin municipal)",
                "operationId": "d88103c18963c6d97662fbc1b78d4777",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Municipalité mise à jour"
                    },
                    "403": {
                        "description": "Admin municipal non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum (format : Bearer <token>)"
            }
        }
    },
...
}